// src/models/account-receive-transfer.ts
export interface AccountReceiveTransfer {
  id: number
  username: string
  email: string
  mobile: string
  id_card: string
  name: string
  last_name: string
  image_url: string
  user_cate_name: string
  address: string
  PROVINCE_NAME: string
  AMPHUR_NAME: string
  DISTRICT_NAME: string
  zipcode: string
  account_saving_id: number
}